added new Species attributes related to CoolProp
For liquid phase reactions, new Species attributes called
isSolventinCoolProp and SolventNameinCoolProp are added. RMG first loads
'inCoolProp' and 'NameinCoolProp' attributes from the RMG-database solvent
library (available in the most updated version of RMG-database).
'inCoolProp' is True if the solvent data are available in CoolProp and
'NameinCoolProp' is the solvent name that can be used in CoolProp.
If the solvent data are unavailable in CoolProp, 'inCoolProp' is False and
'NameinCoolProp' is None.
The Species attrubutes, 'isSolventinCoolProp' and 'SolventNameinCoolProp',
correspond to 'inCoolProp' and 'NameinCoolProp' respectively.