added CoolProp-related attributes and unittest
For liquid phase reactions, new Species attributes called
isSolventinCoolProp and SolventNameinCoolProp are added. RMG first loads
'inCoolProp' and 'NameinCoolProp' from the RMG-database solvent
library (available in the most updated version of RMG-database).
'inCoolProp' is True if the solvent data are available in CoolProp and
'NameinCoolProp' is the solvent name that is recognizable by CoolProp.
If the solvent data are unavailable in CoolProp, 'inCoolProp' is False and
'NameinCoolProp' is None.
If old version of RMG-database is used, then 'inCoolProp' and
'NameinCoolProp' are None by default.
The Species attrubutes, 'isSolventinCoolProp' and 'SolventNameinCoolProp',
correspond to 'inCoolProp' and 'NameinCoolProp' in the database
respectively.