added Thermo correction for the solvent species
For the solvents available in CoolProp, different Thermo correction is
used. Because the solution is very dilute, the solvent thermo can be
assumed to be the same as the pure liquid solvent thermo.
In rmgpy/data/solvation,py, three methods are added: calcSolventH298,
calSolventS298, and getSolventThermo. calcSolventH298 and calSolventS298
compute the standard formation enthalpy and entropy respectively for the
solvent species in liquid phase. These are calculated by using CoolProp
and then adjust its reference enthalpy / entropy to match the ones used in
RMG. Then getSolventThermo obtains the Cpdata for the solvent species from
CoolProp and use them to fit the Wilhoit model, and outputs the new
wilhoit for liquid phase.
This correction is made in rmgpy.rmg.model.py for the solvent species that
are available in CoolProp.

I also fixed a typo in rmgpy.thermo.wilhoit.pyx.fitToData. For fitTodata,
H298 value must be in J/mol not in kJ/mol.