add a def that calculates the solute's K-factor
For solute species, a new def that can compute a continuous and smooth
curve for the K-factor is added. It uses the Abraham LSER, the Mintz LSER,
the linearity of Tln(K-factor) w.r.t. (rho-rho_c) near the critical point
of the solvent, and the critical limit of the K-factor. Tln(K-factor) vs.
(rho-rho_c) is expressed as a 2nd order polynomial for 298 K < T < 380 K
and it switches to a linear function for 380 K < T < T_c.
The curve is smooth and continuous at 380 K.
The details of the new def are explained in rmgpy/data/solvation.py