fix the diffusion limit for rxns involving solvent
If the solvent is involved in the reaction, the direction in which the
solvent reacts is not diffusion-limited.
For example, if the solvent is one of the reactants, the forward rate will
not be diffusion-limited. Or if the solvent is one of the products, the
reverse rate will not be diffusion-limited.