Prevent CoolProp from crashing when T > T_critical
CoolProp crashses when the reactor tempearture is greater than the
critical temperature of the solvent.
Instead of throwing an exception and stopping the rmg run for T >
T_critical, it switches to using the linear temperature dependence of
solvation free energy, which does not require CoolProp, to allow the
users the flexibility of inputting any temperatures they want.