cleaning up solvation-related code
1. RMG() class now has a new class Solvent() that contains solventName,
solventSpecies, and solventData attributes
	solventName: a string for the solventName. the same as the
		     the solvent's name in the database
	solventSpecies: an instance of the class Species(). This is
			the species for the solvent molecule
	solventData: contains the solvent data such as Abraham
		       descriptors, Mintz descriptors and etc. If
		       the new database is used, it also contains
		       2 additional attributes, inCoolProp and
		       nameinCoolProp.
A wrapper class for solvent was created and stored under RMG() class
instead of Species() class because this information is identical across
all species objects.

2. Removed the class Species in rmgpy/rmg/model.py. All the solvation
related attributes will be stored in a new separate class (this new class
is not coded in this commit). another additional attribute,
corSizeAtCreation, is now moved to under rmgpy/species class Species().
All the class Species() import from rmgpy/rmg/model is now replaced by
the import from rmgpy/species

3. Some lines of the code are moved to rmgpy/rmg/input:

  Identifying the solvent from the initial species is now done in
  rmgpy/rmg/input instead of in rmgpy/rmg/main. In rmgpy/rmg/input, it
  identifies the initial species as the solvent by comparting their string
  names. If none of the initial species has the same string name as the
  solvent, it raises the InputError. After identifying the initial species
  as the solvent, it stores that species as solventSpecies attribute in
  class Solvent().

  Storing the indices of the constant concentration species is now doen in
  rmgpy/rmg/input instead of using the get_constSPCIndices method from
  rmgpy/solver/liquid.pyx. In rmgpy/rmg/input, it finds the indices of the
  constant concentration species by comparing the string name of the
  initial species. It stores the indices list under LiquidReactor() class.
  Now, LiquidReactor() class in rmgpy/solvent/liquid does not have
  constSPNames attribute and has constantSpcIndices attribute.

4. In rmgpy/data/solvation, the checkSolventInitialSpecies method is
replaced by checkSolventStructure method. It takes the solvent
class as the input, loads the corresponding solvent structure from the
solvent database by using the getSolventStructure method, and checks
whether the solvent has the correct molecular structure by performing
isIsomorphic test. If the solvent's molecular structure is different from
the one listed in solvent database, it raises DatabaseError.
If the old version of the database is used, the loaded solventStructure is
None, so this test will be skipped.
The checkSolventStructure method is used in rmgpy/rmg/main when it
performs all liquid-phase startup things. solventData is also loaded in
rmgpy/rmg/main.

5. In SolventData() class in rmgpy/data/solvation, two new attributes,
inCoolProp and NameinCoolProp, are added.
	InCoolProp: True if the solvent can be found in the CoolProp
		module. False if not.
	NameinCoolProp: contains the string name of the solvent that is
			used in CoolProp. None if the solvent is not found
			in CoolProp.
These two attributes are None if the old version of the database that does
not contain these information is used.

6. thermoengine code is changed temporarily to avoid error