fix diffusion limit for rxns involving the solvent
When calculating the effective rate, the reaction direction in which the
solvent is involved is not diffusion limited. e.g. If the solvent is one
of the reactants, the forward rate is not diffusion limited. If the
solvent is one of the products, the reverse rate is not diffusion limited.