solvent class is stored under RMG() class
Solvent related data were stored under Species() class previously;
however, this was redundant because the solvent data are identical across
all speices. Instead, Solvent() class is stored under RMG() class in
rmgpy/rmg/input.py. SolventName and SolventSpecies are assigned in
rmgpy/rmg/input.py and solventData are assigined in rmgpy/rmg/main.py.

Checking whether the solvent is listed as one of the initial species is
done in rmgpy/rmg/input.py instead of in main.py. In rmgpy/rmg/input, it
identifies the initial species as the solvent by comparing their string
names. If none of the initial species has the same string name as the
solvent, it raises the InputError. After identifying the initial species
as the solvent, it stores that species as solventSpecies attribute in
class Solvent().