add more accurate solv thermo correction &unittest
Previously, the linear temperature dependence of the solvation free energy
was used for liquid phase species. This commit allows RMG to apply more
accurate solvation free energy correction for solute species if the
solvent can be found in CoolProp. It uses two correlations for the
equilibrium ratio of the solute's vapor and liquid mole fractions in a
dilute binary system (K-factor):
1) Havey's correlation for 298K <= T < Ttransition
2) Japas and Levelt Sengers' correlation for Ttransition <= T < Tc.
The correlation parameters can be determined by using the Abraham LSER
and Mintz LSER combined with two equality constraints to force the
K-factor curve to be smooth and continuous at transition temperature
(Ttransition). K-factor can be coverted into solvation free energy by
using the solvent properties estimated from CoolProp.

For solvent species, the equilibrium partition coefficient can be
directly calculated from CoolProp.

The detailed information will be included in the Liquid Phase System
documentation in the next commit.

If the solvent cannot be found in CoolProp, the temperature dependence of
solvation free energy is assumed linear and only constant free energy
correction is applied for both solute and solvent species.