prevents RMG from using E0 and H0 for liquid rxn
When the solvent is available in CoolProp, RMG fits the NASA model to the
liquid phase free energy. However, because the NASA model is a heat
capacity model, fitting it to free energy results in unreasonable values
of H0 and E0. This commit prevents RMG from using E0 and H0 when the
reaction in liquid phase. Currently, there is no method to estimate E0 and
H0 accurately for liquid phase reactions.