Update the solute data group additivity method
Previously, the solute data group additivity method used only
'abraham' and 'nonacentered' groups in the library. And it used
an algorithm of saturating the lone pairs to radicals, then
saturating with H, finding the groups, and turning back to original
molecules. This algorithm was confusing and inefficient.

Now, the solute data group additivity method uses the groups and
algorithm that are similar to those of the thermo data group
addditivity method. The solute data groups now include: group, ring,
polycyclic, longDistanceInteraction_cyclic,
longDistanceInteraction_noncyclic, and radical.
All new functions or modified functions relevant to the solute data
group additivity follow the same structures of the functions
found in rmgpy/data/thermo.py.