Atomtypes for Phosphorus are created
Atomtypes for Phosphorus are created to support phosphorus compounds in
RMG. A total of 20 atomtypes are added for phosphorus, and most of them
are created by referencing to those of nitrogen because nitrogen and
phosphorus share many similar bonding types. For each atomtype, SMILES
strings of sample compounds that correspond to that atomtype have also
been added as comments.  Atomtype attributes such as 'increment_bond',
'decrement_bond', and etc. are also created by referncing to those of
nitrogen. Notice that phosphorus has several new atomtypes that do not
exist for nitrogen, like P5s, P5d, P5td, etc.